Update ConsoleImpl.cs
Is In/Out/Error redirected properties now use a back in field so a check just need to be done once or when Out/In/Error is changed.

Support for changing In/Out encoding. This actions result on a reset of the properties Out/Error (if not redirected) and In (always).